LPS-54480 Move
