+++ modules/frontend/frontend-js-web/src/com/liferay/frontend/js/web/DefaultPortalWebResources.java
@@ -14,7 +14,7 @@
 
 package com.liferay.frontend.js.web;
 
-import com.liferay.portal.kernel.web.PortalWebResources;
+import com.liferay.portal.kernel.servlet.PortalWebResources;
 
 import javax.servlet.ServletContext;
