--- portal-service/src/com/liferay/portal/kernel/web/packageinfo
@@ -1,1 +1,0 @@
-version 7.0.0
